Updated PublishCloudPrinter
Added functionality to also queries Azure AD Account and gives user the option to select the MCS, HCP and Native App that feeds it into Variables as required parameters. Now you do not have to log in to Azure to go and view what your applications are aclled and youa lso do not need to know the App IDs. ***This script will only work if your DiscoveryEndpoint and DiscoveryResourceID are the same